pyproject.toml
[project]
name = "aventurier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1.7",
    "djangorestframework>=3.16.0",
    "psycopg>=3.2.6",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.2",
]
